Sources Added [Publication] and new module: Analyzer added
* The new sources include Scopus, Science Direct and Microsoft
Academics. Their API KEYS are hidden. These sources won’t work as the
getter does not have the API-Keys. For the keys generate your own or
ask Vishrut (@vreddi).

* New module graphAnalyzer has been added in the controllers. This
provides a visual analysis to the tool to see the spread of researchers
representation over the different aspects and the sources. It works
with all existing and upcoming sources. No new module call needs to be
made to activate the service for graphAnalyzer.